Start update to bevy 0.15
I think the library is ready now, I just need to go through the examples and update them to the new component structure. The main thing I had to adjust is the way Handles are represented in components; while I was there, I deprecated the Bundles and added Component dependencies.